requirements.txt
asgiref==3.3.4
Django==3.0.5
django-compat==1.0.15
djangorestframework==3.12.4
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
